add base class for events (#2)
Note that Event class should store an array of artists rather than 1 artist.
But Jackson mapping of arraylist is ugly, so for now only 1 artist is stored.

Eventfactory should provide methods for all event apis that should be parsed.

that hartl-guy can work in the views with the properties of the Event class.